make weights public
