+++ models.go
@@ -4,30 +4,6 @@ import (
 	"time"
 )
 
-type weights [13]float64
-
-func defaultWeights() weights {
-	return weights{1, 1, 5, -0.5, -0.5, 0.2, 1.4, -0.12, 0.8, 2, -0.2, 0.2, 1}
-}
-
-type Parameters struct {
-	RequestRetention float64
-	MaximumInterval  float64
-	EasyBonus        float64
-	HardFactor       float64
-	W                weights
-}
-
-func DefaultParam() Parameters {
-	return Parameters{
-		RequestRetention: 0.9,
-		MaximumInterval:  36500,
-		EasyBonus:        1.3,
-		HardFactor:       1.2,
-		W:                defaultWeights(),
-	}
-}
-
 type Card struct {
 	Due           time.Time `json:"Due"`
 	Stability     float64   `json:"Stability"`
